### 2021-01-04 #2935 @whedon
[PRE REVIEW]: Thing Directory: Simple and lightweight registry of IoT device metadata
**Submitting author:** @farshidtz (<a href="http://orcid.org/0000-0002-5415-9899">Farshid Tavakolizadeh</a>)
**Repository:** <a href="https://github.com/linksmart/thing-directory" target ="_blank">https://github.com/linksmart/thing-directory</a>
**Version:** v1.0.0-beta.20
