### 2021-01-04 #2935 @arfon
@gkthiruvathukal - would you be willing to edit this submission for JOSS? Also, if you believe this needs a scope check by the JOSS editorial team then we can do that first.

Note, there's a lot of code in `vendor`, removing that and re-running `cloc` gives the following:

```
cloc thing-directory/
      33 text files.
      33 unique files.
       8 files ignored.

Language                     files          blank        comment           code
Go                              17            585            355           2738
JSON                             2              0              0           1086
YAML                             3             46             17            712
TeX                              1              0              0             84
Markdown                         2             25              0             77
Dockerfile                       1             10              2             22
SUM:                            26            666            374           4719
```
