+++ paper/joss/paper.bib
@@ -80,5 +80,6 @@
 	pages        = {131--139},
 	url          = {https://www.eurecom.fr/en/publication/5024/download/comsys-publi-5024.pdf},
 	keywords     = {Discovery,Internet of Things},
-	mendeley-groups = {StorageManagement,StorageManagement/Similar works}
+	mendeley-groups = {StorageManagement,StorageManagement/Similar works},
+	doi          = {10.1145/2991561.2991570}
 }
