### 2021-01-04 #2935 @farshidtz
@arfon - Thanks for the stats. The historical commit information, by author also seems wrong when compared with [contributors on Github](https://github.com/linksmart/thing-directory/graphs/contributors) (the second author is @dschowta).

I have added the missing DOI 10.1145/2991561.2991570 for title: A Categorization of Discovery Technologies for the Internet of Things.
