### 2021-01-04 #2935 @danielskatz
HI @gkthiruvathukal  - It's likely just as easy to write `@whedon assign me as editor` :)
