### 2021-01-19 42b5f6f
Create codeql-analysis.yml
