### 2021-03-07 #3075 @sarathsp06
@farshidtz had a brief look at the article proof and repository
well done! Liked the clear and concise documentation within the code.

I had some minor concerns
* There are some dead code
       * eg: [`enumValues_DataSchemaType` ](https://github.com/linksmart/thing-directory/blob/42b5f6f97c6fb63c89760a01160b89713f6c040b/wot/thing_description.go#L466)

* There are some semantic issues like having the same JSON struct tags for multiple fields in a struct
       * eg: [wot.Link](https://github.com/linksmart/thing-directory/blob/42b5f6f97c6fb63c89760a01160b89713f6c040b/wot/thing_description.go#L202)
* There are quite a few naming convention violations `enumValues_DataSchemaType` is an example and camel cased file names are also not encouraged
* [main.registerDNSSDService](https://github.com/linksmart/thing-directory/blob/42b5f6f97c6fb63c89760a01160b89713f6c040b/discovery.go#L19) does not seem to do what was expected as line 21 instance assignment is ineffective. ie,  `.` in  `conf.DNSSD.Publish.Instance` in effect is not replaces
* there are some instances of err checks missing
* could you please (optionally )  add a how to use section in README
