### 2021-03-07 1a419f1
improve dataschema type comments and enum values variable name
