### 2021-03-07 a3968d9
added missing error handling based on lint result
