### 2021-03-07 4a5dd20
Improve the getting started section
