### 2021-03-07 #3075 @farshidtz
> @farshidtz had a brief look at the article proof and repository
> well done! Liked the clean code and clear and concise documentation within the code.
>
> I had some minor concerns

Thank you very much for your review, @sarathsp06.
I've addressed your concerns. Please find the responses inline:

> * There are some dead code
>   * eg: [`enumValues_DataSchemaType` ](https://github.com/linksmart/thing-directory/blob/42b5f6f97c6fb63c89760a01160b89713f6c040b/wot/thing_description.go#L466)

This is not used by this software, but others who import the package may use it to check if the value of [DataSchema.DataType](https://github.com/linksmart/thing-directory/blob/4a5dd20b04039a9b7253a794813e7a72c70d8d33/wot/thing_description.go#L276) is one of the allowed ones. I've [added comments and moved this part of the code to a better place](https://github.com/linksmart/thing-directory/commit/1a419f199bd536fa697edbc9cb256f6f446c8033) to clarify that.

> * There are some semantic issues like having the same JSON struct tags for multiple fields in a struct
>   * eg: [wot.Link](https://github.com/linksmart/thing-directory/commit/41f19ee32f62716c1bb8b5f24da88cf4b35af358)

I found and [fixed two](https://github.com/linksmart/thing-directory/commit/75afd93fe1d9e74dd72709fdcbe421706559b03f). There are others reported by linting programs, but those are caused by the class relationship coming from WoT Thing Description. I also want to point out that the structs in this file are generated from [this JSON Schema](https://www.w3.org/TR/wot-thing-description/#json-schema-for-validation) and then modified and commented (the JSON Schema is also available [locally](https://github.com/linksmart/thing-directory/blob/4a5dd20b04039a9b7253a794813e7a72c70d8d33/wot/wot_td_schema.json)). Some strange relationships and names come from the generator.

> * There are quite a few naming convention violations `enumValues_DataSchemaType` is an example and camel cased file names are also not encouraged

Fixed [enumValues_DataSchemaType](https://github.com/linksmart/thing-directory/blob/844c55ca311370978a684df815b6f714935b87f4/wot/thing_description.go#L295) the and [TTL](https://github.com/linksmart/thing-directory/commit/844c55ca311370978a684df815b6f714935b87f4) names to match the convention. As mentioned above, the generator of [thing_description.go](https://github.com/linksmart/thing-directory/blob/844c55ca311370978a684df815b6f714935b87f4/wot/thing_description.go) has made some decisions which are not perfect. We correct those whenever we notice any.

> * [main.registerDNSSDService](https://github.com/linksmart/thing-directory/blob/42b5f6f97c6fb63c89760a01160b89713f6c040b/discovery.go#L19) does not seem to do what was expected as line 21 instance assignment is ineffective. ie,  `.` in  `conf.DNSSD.Publish.Instance` in effect is not replaces

The function registers a DNS-SD service as its name suggests. To do so, it has to replace special characters that may come from user configurations. The function does not intend to change the source configuration. I understand the confusion because the config is passed as pointer, but this is solely to avoid copying the object. We'll consider changing it to pass by value.

> * there are some instances of err checks missing

[Added the missing error handling](https://github.com/linksmart/thing-directory/commit/a3968d9e0532d63493eb43628fb3249adc987d8a) for every function call that the linting software could detect.

> * could you please (optionally )  add a how to use section in README

I [improved the getting started section](https://github.com/linksmart/thing-directory/commit/4a5dd20b04039a9b7253a794813e7a72c70d8d33). Does this cover what you expect?

Thanks again for you feedback, it really helps us to improve the project.
