### 2021-03-07 #3075 @sarathsp06
Thank you , I think the following was not clear # probably in my explanation

main.registerDNSSDService:  does not seem to do what was expected as line 21 instance assignment is ineffective. ie, . in conf.DNSSD.Publish.Instance in effect is not replaces

The first replace is ignored as the following line is over writing the same variable
