### 2021-03-07 b469a46
fixed dns-sd instance name escaping
