### 2021-03-07 #3075 @farshidtz
> Thank you , I think the following was not clear # probably in my explanation
>
> main.registerDNSSDService: does not seem to do what was expected as line 21 instance assignment is ineffective. ie, . in conf.DNSSD.Publish.Instance in effect is not replaces
>
> The first replace is ignored as the following line is over writing the same variable

Right, good catch. [Fixed it and added unit tests](https://github.com/linksmart/thing-directory/commit/b469a462704b7c602f1221088ff8d750d72ee418).
