### 2021-03-07 2594c3e
add dedicated search endpoint for jsonpath
