### 2021-03-08 linksmart/thing-directory#37 @yadudoc
Contributing guide
I'm going over the documentation as part of the JOSS review process and it looks like most things are in great shape. The API documentation with Swagger is very nice. I've gotten the application built and launched with docker and also with just go build. At this point, I usually go to the contributing guide, most commonly a separate file (CONTRIBUTING.*) in the project root that covers how to go about testing locally to make sure everything is working fine. Having CI/CD is great, but testing locally is quicker.

While the contrib file could describe many things, I especially find descriptions of the dev workflow and PR process especially useful. I'm tagging this under `Community guidelines` as part of the JOSS review (https://github.com/openjournals/joss-reviews/issues/3075)
