### 2021-03-08 2ac5afc
Add build, run, test instructions
