### 2021-03-08 cc5bb8d
Add build, run, test instructions (#38)
* Add build, run, test instructions

* Copy file local config files
