### 2021-03-08 linksmart/thing-directory#37 @farshidtz
Thanks @yadudoc for your feedback. I agree that a CONTRIBUTING.md file is common, but we have a wiki and having parallel places for documentations may add some confusion.

I [added missing details](https://github.com/linksmart/thing-directory/commit/cc5bb8d7b04105abff7420b7083364fea9ff927c) in regards to build, testing, and execution to the development section of the README. There is also a contributing section there.

Do you think that covers your concerns now?
