### 2021-03-10 linksmart/thing-directory#37 @yadudoc
I believe so. As your project scales, it might be worth considering setting up issue templates to indicate to users how best to report a bug, or specifics regarding a feature request. I'll mark this one closed.
