### 2021-03-10 linksmart/thing-directory#39 @yadudoc
Citation to a more recent survey paper
Based on a quick search on gscholar about surveys on IoT discovery techniques, I see that there's a more recent paper from 2020 here: https://doi.org/10.1016/j.iot.2020.100293 that might be worth considering adding to the related work section.
