### 2021-03-10 linksmart/thing-directory#39 @yadudoc
Related to https://github.com/openjournals/joss-reviews/issues/3075.
