### 2021-03-10 linksmart/thing-directory#40 @yadudoc
Access control for Demo
Currently, the demo page here: https://github.com/linksmart/thing-directory/wiki/Demo appears to be read-only. I have run a few queries from the swagger UI, and that does work. However, the majority of the functionality requires authorization which doesn't appear to be possible as a user. Can you clarify whether the access control section is relevant to the hosted service or if it applies only to the scenario where the user sets up the service?

Related to: https://github.com/openjournals/joss-reviews/issues/3075
