### 2021-03-10 #3075 @yadudoc
Hi @farshidtz, I've looked over the paper, documentation and poked around with a local deployment as well as the swagger UI, and overall I'm very pleased. I'd also echo @sarathsp06's comment that the code documentation is remarkable!

Here are the couple of items that I've raised as issues in the repo:
- [ ] Citation to a more recent survey paper (https://github.com/linksmart/thing-directory/issues/39)
- [ ] Access control for demo. This is mostly so that I could test out the features that require write access. (https://github.com/linksmart/thing-directory/issues/40)
- [X] Fleshing out the contributing guide (https://github.com/linksmart/thing-directory/issues/37)
