### 2021-03-10 linksmart/thing-directory#40 @farshidtz
The demo was setup to support W3C WoT plugfests, first [here](https://github.com/w3c/wot-testing/blob/main/events/2020.06.Online/prototypes/linksmart.md) and continued every 6 months. The data in this demo comes from the plugfests too; see [here](https://github.com/farshidtz/wot-testing-directory-sync). In last month's plugfest, this demo instance was used to [integrate the directory with NodeRED](https://github.com/w3c/wot-testing/blob/main/events/2021.03.Online/Results/2021-03-05-nodered-directory-ktoumura.pdf).

I agree that the authentication section is consufusing without providing credentials or client secret. I [added a note](https://github.com/linksmart/thing-directory/wiki/Demo/_compare/f194572facbeba5d98cc4f876189802c44530154...e14bf32fde96ebee7fcd88024c514a42c11affdb) to hopefully clarify this. As mentioned in the change, the access details are provided to W3C WoT Plugfest participants. The participants are W3C WoT members or guests.
