### 2021-03-10 49fbab6
add function to list all in bytes
