+++ catalog/catalog.go
@@ -47,6 +47,7 @@ type CatalogController interface {
 	patch(id string, d ThingDescription) error
 	delete(id string) error
 	list(page, perPage int) ([]ThingDescription, int, error)
+	listAllBytes() ([]byte, error)
 	filterJSONPath(path string, page, perPage int) ([]interface{}, int, error)
 	filterJSONPathBytes(query string) ([]byte, error)
 	filterXPath(path string, page, perPage int) ([]interface{}, int, error)
@@ -64,6 +65,7 @@ type Storage interface {
 	delete(id string) error
 	get(id string) (ThingDescription, error)
 	list(page, perPage int) ([]ThingDescription, int, error)
+	listAllBytes() ([]byte, error)
 	total() (int, error)
 	iterator() <-chan ThingDescription
 	Close()
