### 2021-03-10 4f55dea
doi and typo fix for geodan
