+++ paper/joss/paper.bib
@@ -27,7 +27,7 @@
 }
 @misc{gost,
 	title        = {GOST (Go-SensorThings)},
-	author       = {Goedan},
+	author       = {Geodan},
 	year         = 2016,
 	journal      = {GitHub repository},
 	publisher    = {GitHub},
@@ -90,7 +90,7 @@
 	journal      = {Internet of Things},
 	volume       = 12,
 	pages        = 100293,
-	doi          = {https://doi.org/10.1016/j.iot.2020.100293},
+	doi          = {10.1016/j.iot.2020.100293},
 	issn         = {2542-6605},
 	url          = {https://www.sciencedirect.com/science/article/pii/S2542660520301256},
 	keywords     = {Internet of things, Resource discovery, Survey of resource discovery, Security, Privacy, CoAP protocol, UPnP protocol, MQTT protocol, AMQP protocol}
