### 2021-03-10 bb08f5e
Joss additional citation (#41)
adding an additional reference in related work
