+++ paper/joss/paper.bib
@@ -27,7 +27,7 @@
 }
 @misc{gost,
 	title        = {GOST (Go-SensorThings)},
-	author       = {Goedan},
+	author       = {Geodan},
 	year         = 2016,
 	journal      = {GitHub repository},
 	publisher    = {GitHub},
@@ -82,4 +82,16 @@
 	keywords     = {Discovery,Internet of Things},
 	mendeley-groups = {StorageManagement,StorageManagement/Similar works},
 	doi          = {10.1145/2991561.2991570}
+}
+@article{DiscoveryReview2020,
+	title        = {Resource discovery techniques in the internet of things: A review},
+	author       = {Kasem Khalil and Khalid Elgazzar and Mohamed Seliem and Magdy Bayoumi},
+	year         = 2020,
+	journal      = {Internet of Things},
+	volume       = 12,
+	pages        = 100293,
+	doi          = {10.1016/j.iot.2020.100293},
+	issn         = {2542-6605},
+	url          = {https://www.sciencedirect.com/science/article/pii/S2542660520301256},
+	keywords     = {Internet of things, Resource discovery, Survey of resource discovery, Security, Privacy, CoAP protocol, UPnP protocol, MQTT protocol, AMQP protocol}
 }
