### 2021-03-10 linksmart/thing-directory#39 @dschowta
Thank you for the pointer. I added this to the related work section. The diff can be seen here:  https://github.com/linksmart/thing-directory/commit/bb08f5efa6cdf2d5b357e2491e306d8f386f8feb
