### 2021-03-12 linksmart/thing-directory#40 @yadudoc
Okay, I think the clarification does help. So, in this case, setting up the service locally with auth, and having a swagger UI point at that for manual testing is probably too much effort. I think I'm going to have to be satisfied with the tests. I'll close this issue.
