### 2021-03-12 linksmart/thing-directory#39 @yadudoc
Thanks @dschowta, I see the changes on the pdf now. Let me close this one.
