### 2021-03-12 linksmart/thing-directory#40 @farshidtz
If you want the the access details, I can send to you by email. But you don't need to worry about the authentication implementation as this is a well tested [library](https://github.com/linksmart/go-sec) used by several other services. It is not a part of this project.
