### 2021-03-12 linksmart/thing-directory#40 @yadudoc
Oh! that would be wonderful. I wanted to try adding a couple of mock devices and query them over the REST API.
