### 2021-03-12 linksmart/thing-directory#40 @farshidtz
Sent you an email. Let me know if you face any issues.
