### 2021-03-13 linksmart/thing-directory#40 @yadudoc
@farshidtz, With the basic auth credentials I was able to register new things and later query them. Thanks for sharing the keys, this definitely made testing out the basic functionality much easier.
