### 2021-03-13 #3075 @yadudoc
Following the discussion over [here](https://github.com/linksmart/thing-directory/issues/40), @farshidtz provided the keys for write access to the demo instance of the thing-directory. I've tested the advertised functionality using the [swagger UI](https://linksmart.github.io/swagger-ui/dist/?url=https://demo.linksmart.eu/thing-directory/openapi-spec-proxy/thing-directory/) and can confirm that everything seems to be in good order. In addition, all of my concerns raised above have been addressed.

@gkthiruvathukal, At this point, I would recommend accepting this paper.
