+++ wot/error.go
@@ -1,0 +1,30 @@
+package wot
+
+// RFC7807 Problem Details (https://tools.ietf.org/html/rfc7807)
+type ProblemDetails struct {
+	// Type A URI reference (RFC3986) that identifies the problem type.  This specification encourages that, when
+	//      dereferenced, it provide human-readable documentation for the problem type (e.g., using HTML).  When
+	//      this member is not present, its value is assumed to be "about:blank".
+	Type string `json:"type,omitempty"`
+
+	// Title - A short, human-readable summary of the problem type. It SHOULD NOT change from occurrence to occurrence of the
+	//      problem, except for purposes of localization (e.g., using proactive content negotiation); see RFC7231, Section 3.4.
+	Title string `json:"title"`
+
+	// Status - The HTTP status code (RFC7231, Section 6) generated by the origin server for this occurrence of the problem.
+	Status int `json:"status"`
+
+	// Detail - A human-readable explanation specific to this occurrence of the problem.
+	Detail string `json:"detail"`
+
+	// Instance - A URI reference that identifies the specific occurrence of the problem.  It may or may not yield further information if dereferenced.
+	Instance string `json:"instance,omitempty"`
+
+	// ValidationErrors - Extension for detailed validation results
+	ValidationErrors []ValidationError `json:"validationErrors,omitempty"`
+}
+
+type ValidationError struct {
+	Name   string `json:"name"`
+	Reason string `json:"reason"`
+}
