+++ catalog/errors.go
@@ -4,9 +4,12 @@ package catalog
 
 import (
 	"encoding/json"
+	"fmt"
 	"log"
 	"net/http"
-	"strings"
+
+	"github.com/linksmart/thing-directory/wot"
+	uuid "github.com/satori/go.uuid"
 )
 
 // Not Found
@@ -24,30 +27,48 @@ type BadRequestError struct{ s string }
 
 func (e *BadRequestError) Error() string { return e.s }
 
-// Error describes an API error (serializable in JSON)
-type Error struct {
-	// Code is the (http) code of the error
-	Code int `json:"code"`
-	// Message is the (human-readable) error message
-	Message string `json:"message"`
+// Validation error (HTTP Bad Request)
+type ValidationError struct {
+	validationErrors []wot.ValidationError
+}
+
+func (e *ValidationError) Error() string { return "validation errors" }
+
+// ErrorResponse writes error to HTTP ResponseWriter
+func ErrorResponse(w http.ResponseWriter, code int, msg ...interface{}) {
+	ProblemDetailsResponse(w, wot.ProblemDetails{
+		Status: code,
+		Detail: fmt.Sprint(msg...),
+	})
+}
+
+func ValidationErrorResponse(w http.ResponseWriter, validationIssues []wot.ValidationError) {
+	ProblemDetailsResponse(w, wot.ProblemDetails{
+		Status:           http.StatusBadRequest,
+		Detail:           "The input did not pass the JSON Schema validation",
+		ValidationErrors: validationIssues,
+	})
 }
 
 // ErrorResponse writes error to HTTP ResponseWriter
-func ErrorResponse(w http.ResponseWriter, code int, msgs ...string) {
-	msg := strings.Join(msgs, " ")
-	e := &Error{
-		code,
-		msg,
+func ProblemDetailsResponse(w http.ResponseWriter, pd wot.ProblemDetails) {
+	if pd.Title == "" {
+		pd.Title = http.StatusText(pd.Status)
+		if pd.Title == "" {
+			panic(fmt.Sprint("Invalid HTTP status code: ", pd.Status))
+		}
 	}
-	if code >= 500 {
-		log.Println("ERROR:", msg)
+	pd.Instance = "/errors/" + uuid.NewV4().String()
+	log.Println("Problem Details instance:", pd.Instance)
+	if pd.Status >= 500 {
+		log.Println("ERROR:", pd.Detail)
 	}
-	b, err := json.Marshal(e)
+	b, err := json.Marshal(pd)
 	if err != nil {
 		log.Printf("ERROR serializing error object: %s", err)
 	}
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(code)
+	w.Header().Set("Content-Type", "application/problem+json")
+	w.WriteHeader(pd.Status)
 	_, err = w.Write(b)
 	if err != nil {
 		log.Printf("ERROR writing HTTP response: %s", err)
