+++ catalog/http.go
@@ -11,7 +11,6 @@ import (
 
 	"github.com/gorilla/mux"
 	"github.com/linksmart/service-catalog/v3/utils"
-	"github.com/linksmart/thing-directory/wot"
 )
 
 const (
@@ -87,6 +86,9 @@ func (a *HTTPAPI) Post(w http.ResponseWriter, req *http.Request) {
 		case *BadRequestError:
 			ErrorResponse(w, http.StatusBadRequest, "Invalid registration:", err.Error())
 			return
+		case *ValidationError:
+			ValidationErrorResponse(w, err.(*ValidationError).validationErrors)
+			return
 		default:
 			ErrorResponse(w, http.StatusInternalServerError, "Error creating the registration:", err.Error())
 			return
@@ -138,6 +140,9 @@ func (a *HTTPAPI) Put(w http.ResponseWriter, req *http.Request) {
 				case *BadRequestError:
 					ErrorResponse(w, http.StatusBadRequest, "Invalid registration:", err.Error())
 					return
+				case *ValidationError:
+					ValidationErrorResponse(w, err.(*ValidationError).validationErrors)
+					return
 				default:
 					ErrorResponse(w, http.StatusInternalServerError, "Error creating the registration:", err.Error())
 					return
@@ -149,6 +154,9 @@ func (a *HTTPAPI) Put(w http.ResponseWriter, req *http.Request) {
 			return
 		case *BadRequestError:
 			ErrorResponse(w, http.StatusBadRequest, "Invalid registration:", err.Error())
+			return
+		case *ValidationError:
+			ValidationErrorResponse(w, err.(*ValidationError).validationErrors)
 			return
 		default:
 			ErrorResponse(w, http.StatusInternalServerError, "Error updating the registration:", err.Error())
@@ -191,6 +199,9 @@ func (a *HTTPAPI) Patch(w http.ResponseWriter, req *http.Request) {
 			return
 		case *BadRequestError:
 			ErrorResponse(w, http.StatusBadRequest, "Invalid registration:", err.Error())
+			return
+		case *ValidationError:
+			ValidationErrorResponse(w, err.(*ValidationError).validationErrors)
 			return
 		default:
 			ErrorResponse(w, http.StatusInternalServerError, "Error updating the registration:", err.Error())
@@ -356,11 +367,14 @@ func (a *HTTPAPI) GetValidation(w http.ResponseWriter, req *http.Request) {
 	}
 
 	var response ValidationResult
-	if err := validateThingDescription(td); err != nil {
-		if verr, ok := err.(*wot.ValidationError); ok {
-			response.Errors = verr.Errors
-		} else {
-			response.Errors = []string{err.Error()}
+	results, err := validateThingDescription(td)
+	if err != nil {
+		ErrorResponse(w, http.StatusInternalServerError, err.Error())
+		return
+	}
+	if len(results) != 0 {
+		for _, result := range results {
+			response.Errors = append(response.Errors, fmt.Sprintf("%s: %s", result.Name, result.Reason))
 		}
 	} else {
 		response.Valid = true
