### 2021-03-14 566753c
add version to swagger spec in SwaggerUI url query param, refactoring
