### 2021-03-15 c8ad11f
Switch to template, set urls as constants
