### 2021-03-15 503b15b
Refactor landing page HTML (#43)
* add version to swagger spec in SwaggerUI url query param, refactoring

* Switch to template, set urls as constants
