### 2021-03-18 #3075 @farshidtz
Thanks a lot to @yadudoc and @sarathsp06 for their thorough feedback.

@gkthiruvathukal, we have made the release and uploaded the archive to Zenodo. Please find the details below:

Version: v1.0.0-beta.21
DOI: https://doi.org/10.5281/zenodo.4616483
