### 2021-03-19 643f029
Add dedicated search endpoints (#36)
* add dedicated search endpoint for jsonpath

* add spec for search api

* add function to list all in bytes

* compliant xpath query

* Increment version
