+++ catalog/controller.go
@@ -159,7 +159,13 @@ func (c *Controller) listAll() ([]ThingDescription, int, error) {
 	}
 }
 
-// TODO: Improve filterJSONPath by reducing the number of (de-)serializations
+func (c *Controller) listAllBytes() ([]byte, error) {
+	return c.storage.listAllBytes()
+}
+
+// Deprecated
+// Note: filterJSONPath performs several (de-)serializations
+// Use filterJSONPathBytes to query bytes directly
 func (c *Controller) filterJSONPath(path string, page, perPage int) ([]interface{}, int, error) {
 	var results []interface{}
 
@@ -201,7 +207,25 @@ func (c *Controller) filterJSONPath(path string, page, perPage int) ([]interface
 	return results[offset : offset+limit], len(results), nil
 }
 
-// TODO: Improve filterXPath by reducing the number of (de-)serializations
+func (c *Controller) filterJSONPathBytes(query string) ([]byte, error) {
+	// query all items
+	b, err := c.listAllBytes()
+	if err != nil {
+		return nil, err
+	}
+
+	// filter results with jsonpath
+	b, err = jsonpath.Get(b, query)
+	if err != nil {
+		return nil, &BadRequestError{fmt.Sprintf("error evaluating jsonpath: %s", err)}
+	}
+
+	return b, nil
+}
+
+// Deprecated
+// Note: filterXPath performs several (de-)serializations
+// Use filterXPathBytes to query bytes directly
 func (c *Controller) filterXPath(path string, page, perPage int) ([]interface{}, int, error) {
 	var results []interface{}
 
@@ -244,6 +268,38 @@ func (c *Controller) filterXPath(path string, page, perPage int) ([]interface{},
 	}
 	// return the requested page
 	return results[offset : offset+limit], len(results), nil
+}
+
+func (c *Controller) filterXPathBytes(path string) ([]byte, error) {
+	// query all items
+	b, err := c.listAllBytes()
+	if err != nil {
+		return nil, err
+	}
+
+	// parse the json document
+	doc, err := xpath.Parse(bytes.NewReader(b))
+	if err != nil {
+		return nil, fmt.Errorf("error parsing serialized input for xpath filtering: %s", err)
+	}
+
+	// filter with xpath
+	nodes, err := xpath.QueryAll(doc, path)
+	if err != nil {
+		return nil, &BadRequestError{fmt.Sprintf("error filtering input with xpath: %s", err)}
+	}
+	results := make([]interface{}, len(nodes))
+	for i := range nodes {
+		results[i] = getObjectFromXPathNode(nodes[i])
+	}
+
+	// serialize
+	b, err = json.Marshal(results)
+	if err != nil {
+		return nil, fmt.Errorf("error serliazing results of xpath filtering: %s", err)
+	}
+
+	return b, nil
 }
 
 // basicTypeFromXPathStr is a hack to get the actual data type from xpath.TextNode
