### 2021-03-19 f630e81
remove space in logo, add logo to http root endpoint response
