### 2021-03-24 8aa7265
remove delay in response encoding
