### 2021-04-14 #3075 @kyleniemeyer
Hi @farshidtz, I am the AEIC on duty this week doing final checks before accepting. I noticed that one reference (Schaaf 2016, for the FROST server) seems to have an issue with the author names. I think if you change the `author` field in the `.bib` file to `{Hylke {van der Schaaf} and Michael Jacoby}` or `{van der Schaaf, Hylke and Jacoby, Michael}` then bibtex should parse it correctly.
