+++ paper/joss/paper.bib
@@ -19,7 +19,7 @@
 }
 @misc{frost,
 	title        = {FROST Server},
-	author       = {van der Schaaf and H., Jacoby, M},
+	author       = {{van der Schaaf}, Hylke and Jacoby, Michael},
 	year         = 2016,
 	journal      = {GitHub repository},
 	publisher    = {GitHub},
