### 2021-04-14 #3075 @farshidtz
Thanks @kyleniemeyer. I have changed it and the produced [artifact](https://github.com/linksmart/thing-directory/suites/2496106542/artifacts/53883249) shows the name correctly now.
