### 2021-04-14 #3075 @kyleniemeyer
Many thanks to @yadudoc and @sarathsp06 for their reviews, and @gkthiruvathukal for editing this.

It looks like the DOI is not yet resolving, so I'm going to leave this issue open until that step is finished. You can see the final paper here: https://joss.theoj.org/papers/10.21105/joss.03075
